+++ models/default-model/package.json
@@ -29,10 +29,10 @@
     "lint": "pnpm scaffold:dependencies && eslint -c ../.eslintrc.cjs src --ext .ts",
     "prepublishOnly": "pnpm lint && pnpm build && pnpm validate:build",
     "validate:build": "ts-node ../../scripts/package-scripts/validate-build.ts models/default-model",
-    "build": "ts-node ../../scripts/package-scripts/build-model.ts default-model -o cjs -o esm -o umd --skipCheckModelsExist",
-    "build:cjs": "ts-node ../../scripts/package-scripts/build-model.ts default-model -o cjs",
-    "build:esm": "ts-node ../../scripts/package-scripts/build-model.ts default-model -o esm",
-    "build:umd": "ts-node ../../scripts/package-scripts/build-model.ts default-model -o umd",
+    "build": "ts-node ../../scripts/package-scripts/build-model.cts default-model -o cjs -o esm -o umd --skipCheckModelsExist",
+    "build:cjs": "ts-node ../../scripts/package-scripts/build-model.cts default-model -o cjs",
+    "build:esm": "ts-node ../../scripts/package-scripts/build-model.cts default-model -o esm",
+    "build:umd": "ts-node ../../scripts/package-scripts/build-model.cts default-model -o umd",
     "clean": "ts-node ../../scripts/package-scripts/clean-model.ts default-model"
   },
   "files": [
