+++ models/esrgan-experiments/package.json
@@ -4222,7 +4222,7 @@
     "lint": "pnpm scaffold:dependencies && eslint -c ../.eslintrc.cjs src --ext .ts",
     "prepublishOnly": "pnpm lint && pnpm build && pnpm validate:build",
     "validate:build": "ts-node ../../scripts/package-scripts/validate-build.ts models/esrgan-experiments",
-    "build": "ts-node ../../scripts/package-scripts/build-model.ts esrgan-experiments -o cjs -o esm -o umd"
+    "build": "ts-node ../../scripts/package-scripts/build-model.cts esrgan-experiments -o cjs -o esm -o umd"
   },
   "keywords": [
     "image super resolution",
