+++ models/esrgan-legacy/package.json
@@ -49,10 +49,10 @@
     "lint": "pnpm scaffold:dependencies && eslint -c ../.eslintrc.cjs src --ext .ts",
     "prepublishOnly": "pnpm lint && pnpm build && pnpm validate:build",
     "validate:build": "ts-node ../../scripts/package-scripts/validate-build.ts models/esrgan-legacy",
-    "build": "ts-node ../../scripts/package-scripts/build-model.ts esrgan-legacy -o cjs -o esm -o umd",
-    "build:cjs": "ts-node ../../scripts/package-scripts/build-model.ts esrgan-legacy -o cjs",
-    "build:esm": "ts-node ../../scripts/package-scripts/build-model.ts esrgan-legacy -o esm",
-    "build:umd": "ts-node ../../scripts/package-scripts/build-model.ts esrgan-legacy -o umd",
+    "build": "ts-node ../../scripts/package-scripts/build-model.cts esrgan-legacy -o cjs -o esm -o umd",
+    "build:cjs": "ts-node ../../scripts/package-scripts/build-model.cts esrgan-legacy -o cjs",
+    "build:esm": "ts-node ../../scripts/package-scripts/build-model.cts esrgan-legacy -o esm",
+    "build:umd": "ts-node ../../scripts/package-scripts/build-model.cts esrgan-legacy -o umd",
     "clean": "ts-node ../../scripts/package-scripts/clean-model.ts esrgan-legacy"
   },
   "files": [
