+++ models/esrgan-medium/package.json
@@ -45,10 +45,10 @@
     "lint": "pnpm scaffold:dependencies && eslint -c ../.eslintrc.cjs src --ext .ts",
     "prepublishOnly": "pnpm lint && pnpm build && pnpm validate:build",
     "validate:build": "ts-node ../../scripts/package-scripts/validate-build.ts models/esrgan-medium",
-    "build": "ts-node ../../scripts/package-scripts/build-model.ts esrgan-medium -o cjs -o esm -o umd",
-    "build:cjs": "ts-node ../../scripts/package-scripts/build-model.ts esrgan-medium -o cjs",
-    "build:esm": "ts-node ../../scripts/package-scripts/build-model.ts esrgan-medium -o esm",
-    "build:umd": "ts-node ../../scripts/package-scripts/build-model.ts esrgan-medium -o umd",
+    "build": "ts-node ../../scripts/package-scripts/build-model.cts esrgan-medium -o cjs -o esm -o umd",
+    "build:cjs": "ts-node ../../scripts/package-scripts/build-model.cts esrgan-medium -o cjs",
+    "build:esm": "ts-node ../../scripts/package-scripts/build-model.cts esrgan-medium -o esm",
+    "build:umd": "ts-node ../../scripts/package-scripts/build-model.cts esrgan-medium -o umd",
     "clean": "ts-node ../../scripts/package-scripts/clean-model.ts esrgan-medium"
   },
   "files": [
