+++ scripts/package-scripts/build-core.cts
