+++ scripts/package-scripts/build-model.cts
